Extract Router::Candidates (verbatim move, no behavior change)
NxN G14 3c, R12: code-move only. Move the candidate-selection,
exclusion, and scoring helpers (select_candidates, the seven
excluded_by_* filters, satisfies_required_capabilities?, pick_best,
effective_priority, hint_matching_bonus, tier_priority_bonus,
effort_matching_bonus, effort_for_tier, loaded_model_bonus) out of
Router into a Router::Candidates mixin, extended into the Router
singleton.

Module is lexically nested under Router so EFFORT_RANK / Discovery /
Legion::Settings resolve identically; extend keeps every method on the
same self with private visibility, so cross-calls to surviving helpers
(tier_available?, tier_rank, tier_priority, health_tracker,
discovery_enabled?, required_capabilities, normalize_capabilities,
normalize_effort, external_tier?) and the @last_candidate_trace ivar are
unchanged. No external caller references these private methods. Router
shrinks from ~1030 to ~660 lines. Full rspec 3219/0, rubocop 490/0.
